:recycle: Begin refactor of diff/change reports
