+++ reports/summary_report.py
@@ -11,8 +11,24 @@ from collections import defaultdict
 
 from jinja2 import Environment, FileSystemLoader
 
+TABLES = [
+    'study',
+    'participant',
+    'biospecimen',
+    'phenotype',
+    'diagnosis',
+    'genomic_file'
+]
+
+IGNORE_COLS = ['uuid', 'created_at', 'modified_at', 'kf_id']
+
 
 def get_pg_connection_str():
+    """
+    Helper function to load database creds from vault
+    """
+    if 'CONN_STR' in os.enviorn:
+        return os.environ['CONN_STR']
     env = os.environ.get('ENV')
     vault_url = os.environ.get('VAULT_URL')
     path = os.environ.get('DATASERVICE_PG_SECRET')
@@ -30,17 +46,6 @@ def get_pg_connection_str():
     host = secret['hostname']
     return f"postgres://{usr}:{pas}@{host}:5432/kfpostgres{env}"
 
-TABLES = [
-    'study',
-    'participant',
-    'biospecimen',
-    'phenotype',
-    'diagnosis',
-    'genomic_file'
-]
-IGNORE_COLS = ['uuid', 'created_at', 'modified_at', 'kf_id']
-DIFF_RE = re.compile(r'^(.*\()([+-]?\d+(\.\d+)?)\)$')
-
 
 def handler(event, context):
     """
@@ -51,42 +56,33 @@ def handler(event, context):
     conn_str = get_pg_connection_str()
 
     g = SummaryGenerator(output=f"/tmp/", s3output=output, conn_str=conn_str)
-    diff_message = g.make_report()
 
-    # Collect all files in /tmp
+    # Collect all files in /tmp for upload
     files = set(glob.glob('/tmp/**/*.png', recursive=True))
     files = files.union(set(glob.glob('/tmp/**/*.csv', recursive=True)))
     files = files.union(set(glob.glob('/tmp/**/*.html', recursive=True)))
 
-    if len(diff_message) > 0:
-        url = event['output']
-        url += f"/Table_Diff_Report_{datetime.now().strftime('%Y%m%d')}.html"
-        url = 'https://s3.amazonaws.com/' + url
-        diff_message[0]["actions"] = [
-            {
-                "type": "button",
-                "text": ":eyes: Check it Out :eyes:",
-                "url": url,
-                "style": "primary"
-            }
-        ]
-
-    return diff_message, {p.replace('/tmp/', ''): p for p in list(files)}
+    return [], {p.replace('/tmp/', ''): p for p in list(files)}
 
 
 class SummaryGenerator:
 
-    def __init__(self, study_id=None, output='', s3output='', conn_str=''):
+    def __init__(self, study_id=None, output='', conn_str=''):
+        """
+        :param study_id: The kf_id of the study to generate a summary for. If
+            `None`, the report will be run for all data.
+        :param output: The path to save results to. If run for a specific
+            study, results will be saved within a directory named by the kf_id
+            of that study in this path.
+        :param conn_str: The sql connection string for the database
+        """
         self.study_id = study_id
         self.conn_str = conn_str
         self.output = output
-        self.s3output = s3output
         if self.study_id is not None:
             self.output += self.study_id+'/'
 
         os.makedirs(self.output+'tables', exist_ok=True)
-        os.makedirs(self.output+'yesterday', exist_ok=True)
-        os.makedirs(self.output+'diffs', exist_ok=True)
 
     def make_report(self):
         """
@@ -115,56 +111,14 @@ class SummaryGenerator:
             "date": datetime.now(),
             "sections": table_summaries
         }
+
         # Render and save HTML
         html_out = template.render(template_vars)
         filename = f"Table_Summary_Report_{datetime.now().strftime('%Y%m%d')}.html"
         with open(self.output+filename, "w+") as f:
             f.write(html_out)
 
-        diff_message = self.make_diff_report(table_summaries)
-
-        return diff_message
-
-    def make_diff_report(self, table_summaries):
-        """
-        Download all of yesterday's summary tables then build the diff report
-        Derive yesterdays s3 location based on the output for this report
-        """
-        m = re.match(r'.*/(\d{8})-reports/.*', self.s3output)
-        if not m or not m.groups(1) or len(m.groups(1)) == 0:
-            print(f"yesterday's date could not be resolved: {m}")
-            return
-        yesterday = datetime.strptime(m.groups(1)[0], '%Y%m%d') - timedelta(days=1)
-        yesterday = datetime.strftime(yesterday, '%Y%m%d')
-        yesterday_path = self.s3output.replace(m.groups(1)[0], yesterday)
-        yesterday_path = yesterday_path + '/tables/'
-        bucket = yesterday_path.split('/')[0]
-        yesterday_path = '/'.join(yesterday_path.split('/')[1:])
-
-        client = boto3.client('s3')
-        paginator = client.get_paginator('list_objects')
-        print(f'Downloading all summaries from {bucket}{yesterday_path}')
-
-        for page in paginator.paginate(Bucket=bucket, Prefix=yesterday_path):
-            if 'Contents' not in page or len(page['Contents']) == 0:
-                print('Could not find data from yesterday')
-                return
-            for obj in page['Contents']:
-                fname = obj['Key'].split('/')[-1]
-                client.download_file(bucket,
-                                     obj['Key'],
-                                     self.output+'yesterday/'+fname)
-
-
-        diff = DiffGenerator(output=self.output+'diffs/',
-                             summaries=table_summaries,
-                             yesterday=self.output+'yesterday/')
-        message = diff.make_report()
-
-        # Remove yesterday's data so it's not uploaded again
-        import shutil
-        shutil.rmtree(self.output+'yesterday')
-        return message
+        # TODO Call diff report lambda
 
 def table_report(df):
     """
@@ -196,228 +150,6 @@ def column_report(df, col):
                        .reset_index()
                        .rename(columns={'index':'count'}))
     return counts
-
-
-class DiffGenerator:
-
-    def __init__(self, output='/tmp/', yesterday=None, summaries=None):
-        """
-        :param output: The directory to output all diff tables
-        :param summaries: The summary tables for the current state
-        :param yesterday: A directory where to find all of yesterday's summaries
-        """
-        if not os.path.isdir(yesterday):
-            raise Exception(f'Please provide valid directory paths')
-
-        self.output = output
-        self.yesterday_path = yesterday
-        self.yesterday_files = os.listdir(yesterday)
-
-        summaries = {f'{k1}_{k2}': v2 for k1, v1 in summaries.items()
-                                     for k2, v2 in v1.items()}
-        self.summaries = summaries
-
-    def make_report(self):
-        diffs = self.compute_diffs()
-
-        env = Environment(loader=FileSystemLoader(os.path.dirname(os.path.abspath(__file__))))
-        template = env.get_template("summary_template.html")
-
-        template_vars = {
-            "title": "Daily Change Report",
-            "date": datetime.now(),
-            "sections": diffs
-        }
-        # Render and save HTML
-        html_out = template.render(template_vars)
-        filename = f"../Table_Diff_Report_{datetime.now().strftime('%Y%m%d')}.html"
-        with open(self.output+filename, "w+") as f:
-            f.write(html_out)
-
-        # Create diff message
-        return self.diff_summary_message()
-
-    def diff_summary_message(self):
-        """
-        Compute number of changes by table then report number of tables
-        changed and number of values changed
-        """
-        counts = self.diff_counts
-        summaries = {f'{k1}_{k2}': v2 for k1, v1 in counts.items()
-                                     for k2, v2 in v1.items()}
-
-        total = int(sum(summaries.values()))
-        non_zero = [k for k, v in summaries.items() if v > 0]
-
-        if total <= 0:
-            return []
-        
-        expl = np.random.choice(['Wow', 'Wowza', 'Woa', 'Gee', 'Oh my', 'Hey',
-                                 'Take a look', 'Check it out', 'Gosh', 'Golly'])
-
-        message = (f"{expl}! There were *{total}* changes made across "
-                    +f"*{len(non_zero)}/{len(summaries)}* columns yesterday!")
-
-        attachments = [{
-            "text": message,
-            "fallback": message,
-            "color": "#3AA3E3",
-            "attachment_type": "default",
-        }]
-
-        return attachments
-
-    def compute_diffs(self):
-        """
-        Looks through summary tables and computes diffs for each returning a
-        dict of styled html tables organized by table and column and leaving
-        out any summaries that did not change.
-        """
-
-        sections = defaultdict(dict)
-        # Also keep track of total number of changes per column
-        counts = defaultdict(dict)
-        # Iterate through the summaries from the summary report and try to
-        # diff against the same summary from yesterday, if it exists
-        for f, df1 in self.summaries.items():
-            section = f.split('_')[0]
-            name = f[f.find('_')+1:].replace('.csv', '')
-
-            df2_path = os.path.join(self.yesterday_path, f+'.csv')
-            # Don't try to diff if it's a summary that didn't exist yesterday
-            if not os.path.exists(df2_path):
-                continue
-            df2 = pd.read_csv(df2_path, index_col=0)
-
-            # Diff using k: v dictonary comparison
-            if len(df1.columns) == 2 and df1.columns[1] == 'count':
-                sections[section][name] = self.count_diff(df1, df2)
-                sections[section][name].to_csv(f'{self.output}{section}_{name}_diff.csv')
-                counts[section][name] = sections[section][name]['change'].abs().sum()
-                sections[section][name] = sections[section][name]\
-                        .drop(['count', 'change', 'summary', 'count_yesterday'],
-                                             axis=1)\
-                        .rename(columns={'summary_html': 'count (change)'})
-            # Diff for wide tables with strings
-            else:
-                sections[section][name] = self.compute_diff(df1, df2)
-                sections[section][name].to_csv(f'{self.output}{section}_{name}_diff.csv')
-            # Don't try to style an empty diff and just return None
-            if sections[section][name].shape[0] == 0:
-                sections[section][name] = None
-                continue
-            if len(df1.columns) == 2 and df1.columns[1] == 'count':
-                sections[section][name] = sections[section][name].style.set_table_attributes('class="table table-striped table-hover"')
-            else:
-              sections[section][name] = sections[section][name].applymap(self.highlight_diff)
-              sections[section][name] = sections[section][name].style.set_table_attributes('class="table table-striped table-hover"')
-
-        self.diff_counts = counts
-        return sections
-
-    def count_diff(self, df1, df2):
-        """
-        Compares two dataframes by converting to dicts
-        """
-        def diff_format(r):
-            return f"{int(r['count'])} ({int(r['change']):+})"
-
-        def diff_html(r):
-            """
-            Wraps values in span tags for display in web page
-            """
-            change = f"{int(r['change']):+}"
-            color = ''
-            if r['change'] > 0:
-                color = 'text-success'
-            elif r['change'] < 0:
-                color = 'text-danger'
-            else:
-                color = 'text-color'
-            change = f'(<span class="{color}">{change}</span>)'
-            return f"{int(r['count'])} {change}"
-
-        # Convert datetimes to strings
-        for c in df1.select_dtypes(include=[np.datetime64]):
-            df1[c] = df1[c].dt.strftime('%Y-%m-%d %H:%M:%S')
-
-        for c in df2.select_dtypes(include=[np.datetime64]):
-            df2[c] = df2[c].dt.strftime('%Y-%m-%d %H:%M:%S')
-
-        diff = df2.merge(df1, how='outer', on=df1.columns[0],
-                         suffixes=['_yesterday', '']).fillna(0)
-        diff['change'] = diff['count'] - diff['count_yesterday']
-        diff = diff.sort_values(['change', 'count'], ascending=False).reset_index(drop=True)
-        diff['summary'] = diff.apply(diff_format, axis=1)
-        diff['summary_html'] = diff.apply(diff_html, axis=1)
-        diff = diff[diff['change'] != 0]
-        diff = diff.reset_index(drop=True)
-        return diff
-
-    def compute_diff(self, df1, df2):
-        """
-        Compares two dataframes
-        """
-        df = pd.concat([df1, df2], 
-                        axis='columns',
-                        keys=['DF1', 'DF2'])
-
-        diff = pd.DataFrame(columns=set(df.columns.get_level_values(1)))
-
-        for c in set(df.columns.get_level_values(1)):
-            if c not in df['DF1'].columns:
-                diff[c] = df['DF2'][c]
-            elif c not in df['DF2'].columns:
-                diff[c] = df['DF1'][c]
-            else:
-                diff[c] = df.xs(c, level=1, drop_level=1, axis=1) \
-                            .apply(self.differ, axis=1)
-            
-        return diff
-
-    def highlight_diff(self, value):
-        if value is None:
-            return value
-        
-        try:
-            m = DIFF_RE.match(str(value))
-            if m:
-                change = float(str(m.group(2)))
-            else:
-                return value
-
-            if change < 0:
-                r = DIFF_RE.sub(r"\1<span class='text-danger'>\2</span>)", value)
-                return r
-            elif change > 0:
-                r = DIFF_RE.sub(r"\1<span class='text-success'>\2</span>)", value)
-                return r
-            else:
-                return value
-        except ValueError as verr:
-            return value
-        return value
-
-    def differ(self, r):
-        """
-        Compares two values and returns the new value
-        """
-        first = r['DF1']
-        second = r['DF2']
-        try:
-            try:
-                first = int(str(first))
-                second = int(str(second))
-            except ValueError:
-                first = float(str(first))
-                second = float(str(second))
-        except ValueError as verr:
-            if str(first) != str(second):
-                return f'<span class="text-success">{first}</span> <span class="text-danger">(-{second})</span>'
-            else:
-                return first
-        return f'{first} ({(first - second):+})'
-
 
 
 # For local testing
